search algos

+++ 0x1E-search_algorithms/0-linear.c
@@ -6,7 +6,6 @@
  * @array: array to be searched
  * @size: size of array
  * @value: value being searched for
- *
  * Return: Always EXIT_SUCCESS
  */
 int linear_search(int *array, size_t size, int value)
@@ -16,7 +15,7 @@ int linear_search(int *array, size_t size, int value)
 	if (array == NULL)
 		return (-1);
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < (size_t)size; i++)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (value == array[i])
